+++ src/components/listView/IListView.ts
@@ -1,4 +1,4 @@
-import { IBook } from '../../pages/mainPage/types';
+import IBook from '../../interfaces/types';
 
 export interface IListView {
   bookList: IBook[];
